Revert "use ms for metrics"
This reverts commit 70f6a5e638d97a26e5faa06ec60d82eda321502a.

+++ dtnsim/src/utils/MetricCollector.h
@@ -61,7 +61,7 @@ class MetricCollector {
     void updateRUCoPComputationTime(long computationTime);
     void updateCGRComputationTime(long computationTime);
     void setNumberOfHops(long bundleId, int hops);
-    void updateBundleElapsedTime(long bundleId, int64_t elapsedTime);
+    void updateBundleElapsedTime(long bundleId, double elapsedTime);
     void intializeArrivalTime(long bundleId, std::chrono::steady_clock::time_point initialTime);
     void setFinalArrivalTime(long bundleId, std::chrono::steady_clock::time_point finalTime);
     void evaluateAndPrintResults();
@@ -93,7 +93,7 @@ class MetricCollector {
     int nodesNumber_;
 
     map<long, int> bundleHops_;
-    map<long, int64_t> bundleElapsedTime_;
+    map<long, double> bundleElapsedTime_;
     map<long, ArrivalInfo> bundleArrivalTime_;
 };
